requirements.txt
flask
line-bot-sdk
google-generativeai
python-dotenv
gunicorn
